django/requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.2.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
docopt==0.6.2
dulwich==0.21.6
filelock==3.12.4
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.17.3
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3.post1
pywin32-ctypes==0.2.2
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.3
sqlparse==0.4.4
tomlkit==0.12.1
trove-classifiers==2023.8.7
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.5
yarg==0.1.9
zipp==3.16.2
